+++ backend/models/course_ucore.py
@@ -3,9 +3,9 @@ from datetime import datetime
 
 
 @dataclass(frozen=True)
-class CourseUcore:
-    id: int
-    course_id: int
-    ucore_id: int
-    created_at: datetime
-    updated_at: datetime
+class CourseUCore:
+    course_id: str  # e.g., HIST 105
+    ucore_designation: str  # e.g., ROOTS
+    course_name: str
+    credits: str
+    created_at: datetime = datetime.now()
